\name{cncaGUI-package}
\alias{cncaGUI-package}
\docType{package}
\title{Canonical Non-symmetrical Correspondence Analysis package}
\description{Provides a GUI with which users can construct and interact
        with Canonical Non-symmetrical Correspondence Analysis}
\details{
\tabular{ll}{
Package: \tab cncaGUI\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2010-07-20\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{Ana Belen Nieto Librero \email{ananieto@usal.es}, 
Priscila Willems \email{pwillems@bariloche.inta.gov.ar},
Purificacion Galindo Villardon \email{pgalindo@usal.es}}

\keyword{ package }
