\name{especies}
\alias{especies}
\docType{data}
\title{Species data
}
\description{
12 species of hunting spiders caught in pitfall
traps in a Dutch dune area Ter-Braak (1986).
}
\usage{data(especies)}
\format{
  A data frame with 28 observations on 12 variables.
}

\references{ Ter-Braak CJF (1986). \dQuote{Canonical Correspondence Analysis: a New Eigenvector Technique
for Multivariate Direct Gradient Analysis.} \emph{Ecology}, \bold{67(5)}, 1167-1179
}
\examples{
data(especies)
}
\keyword{datasets}
