\name{variables}
\alias{variables}
\docType{data}
\title{
Environmental variables data
}
\description{
Six environmental variables measured in 28 sites}
\usage{data(variables)}
\format{
  A data frame with 28 observations on 6 variables.
}
\references{ Ter-Braak CJF (1986). \dQuote{Canonical Correspondence Analysis: a New Eigenvector Technique
for Multivariate Direct Gradient Analysis.} \emph{Ecology}, \bold{67(5)}, 1167-1179
}
\examples{
data(variables)
}
\keyword{datasets}
